fix: clear frontend lint for workflow prompt UI
Extract user-message body helpers so chat-message stays under max-lines,
and de-duplicate the workflow-prompt-section testid literal.
