+++ apps/web/components/task/chat/messages/chat-message.tsx
@@ -10,13 +10,7 @@ import {
   type ReactNode,
 } from "react";
 import type { Components } from "react-markdown";
-import {
-  IconWand,
-  IconMessageDots,
-  IconFile,
-  IconFolder,
-  IconChevronRight,
-} from "@tabler/icons-react";
+import { IconWand, IconMessageDots, IconFile, IconFolder } from "@tabler/icons-react";
 import { useTranslation } from "react-i18next";
 import { cn } from "@/lib/utils";
 import type { Message } from "@/lib/types/http";
@@ -24,7 +18,6 @@ import { MessageActions } from "@/components/task/chat/messages/message-actions"
 import { useMessageFavorite } from "@/hooks/domains/session/use-message-favorite";
 import { useUserMessageNavigation } from "@/hooks/use-message-navigation";
 import { SenderTaskBadge, type SenderTaskInfo } from "./sender-task-badge";
-import { MemoizedMarkdown } from "@/components/shared/memoized-markdown";
 import { ImagePreviewDialog } from "@/components/task/chat/image-preview-dialog";
 import { useAppStore } from "@/components/state-provider";
 import { HoverCard, HoverCardTrigger, HoverCardContent } from "@kandev/ui/hover-card";
@@ -45,9 +38,7 @@ import { entityReferencesFromMetadata } from "@/lib/entity-references/message-re
 import type { EntityReference } from "@/lib/types/entity-reference";
 import { attachmentContentUrl } from "@/lib/api/domains/attachment-api";
 import { formatBytes } from "@/lib/utils/format-bytes";
-import { t } from "@/lib/i18n";
-import { splitWorkflowInstructions } from "@/lib/utils/workflow-instructions";
-import { Collapsible, CollapsibleContent, CollapsibleTrigger } from "@kandev/ui/collapsible";
+import { renderUserMessageBody } from "./user-message-body";
 
 type ChatMessageProps = {
   comment: Message;
@@ -101,144 +92,6 @@ function renderContentWithFileRefs(content: string): React.ReactNode[] {
   }
 
   return parts.length > 0 ? parts : [content];
-}
-
-// ── Markdown component overrides imported from shared/markdown-components ─────
-
-type UserMessageBodyOptions = {
-  hasContent: boolean;
-  showRaw: boolean;
-  hasAttachments: boolean;
-  content: string;
-  rawContent?: string;
-  promptMentionComponents?: Components;
-  taskId: string;
-  worktreePath?: string;
-  onOpenFile?: (path: string) => void;
-};
-
-function UserMessageMarkdown({
-  content,
-  promptMentionComponents,
-  taskId,
-  worktreePath,
-  onOpenFile,
-}: {
-  content: string;
-  promptMentionComponents?: Components;
-  taskId: string;
-  worktreePath?: string;
-  onOpenFile?: (path: string) => void;
-}) {
-  return (
-    <div className="markdown-body markdown-body-user max-w-none">
-      <MemoizedMarkdown
-        content={content}
-        taskId={taskId}
-        components={promptMentionComponents}
-        worktreePath={worktreePath}
-        onOpenFile={onOpenFile}
-      />
-    </div>
-  );
-}
-
-function CollapsedWorkflowInstructions({
-  instructions,
-  rest,
-  promptMentionComponents,
-  taskId,
-  worktreePath,
-  onOpenFile,
-}: {
-  instructions: string;
-  rest: string;
-  promptMentionComponents?: Components;
-  taskId: string;
-  worktreePath?: string;
-  onOpenFile?: (path: string) => void;
-}) {
-  const { t } = useTranslation();
-  const [open, setOpen] = useState(false);
-  return (
-    <div className="space-y-2">
-      <Collapsible open={open} onOpenChange={setOpen}>
-        <CollapsibleTrigger
-          className="flex w-full cursor-pointer items-center gap-1 rounded-md bg-muted/40 px-2 py-1 text-left text-xs text-muted-foreground hover:bg-muted/60"
-          data-testid="workflow-instructions-toggle"
-        >
-          <IconChevronRight
-            className={cn("h-3.5 w-3.5 shrink-0 transition-transform", open && "rotate-90")}
-          />
-          <span>{t("workflows:workflowInstructionsCollapsed")}</span>
-        </CollapsibleTrigger>
-        <CollapsibleContent className="pt-2">
-          <UserMessageMarkdown
-            content={instructions}
-            promptMentionComponents={promptMentionComponents}
-            taskId={taskId}
-            worktreePath={worktreePath}
-            onOpenFile={onOpenFile}
-          />
-        </CollapsibleContent>
-      </Collapsible>
-      {rest.trim() !== "" && (
-        <UserMessageMarkdown
-          content={rest}
-          promptMentionComponents={promptMentionComponents}
-          taskId={taskId}
-          worktreePath={worktreePath}
-          onOpenFile={onOpenFile}
-        />
-      )}
-    </div>
-  );
-}
-
-function renderUserMessageBody({
-  hasContent,
-  showRaw,
-  hasAttachments,
-  content,
-  rawContent,
-  promptMentionComponents,
-  taskId,
-  worktreePath,
-  onOpenFile,
-}: UserMessageBodyOptions): React.ReactNode {
-  if (hasContent && showRaw) {
-    return <pre className="whitespace-pre-wrap font-mono text-xs">{rawContent || content}</pre>;
-  }
-  if (hasContent) {
-    const split = splitWorkflowInstructions(content);
-    if (split.hasInstructions) {
-      return (
-        <CollapsedWorkflowInstructions
-          instructions={split.instructions}
-          rest={split.rest}
-          promptMentionComponents={promptMentionComponents}
-          taskId={taskId}
-          worktreePath={worktreePath}
-          onOpenFile={onOpenFile}
-        />
-      );
-    }
-    return (
-      <UserMessageMarkdown
-        content={content}
-        promptMentionComponents={promptMentionComponents}
-        taskId={taskId}
-        worktreePath={worktreePath}
-        onOpenFile={onOpenFile}
-      />
-    );
-  }
-  if (!hasAttachments) {
-    return (
-      <p className="whitespace-pre-wrap break-words overflow-wrap-anywhere">{t("task:empty")}</p>
-    );
-  }
-  return null;
 }
 
 // ── User message sub-component ──────────────────────────────────────
